like and unlike

+++ controllers/fileController.js
@@ -40,7 +40,11 @@ exports.uploadFile = tryCatch(async (req, res, next) => {
 exports.deleteFile = tryCatch(async (req, res, next) => {
   const { fileName } = req.params;
   const filePath = path.join(__dirname, "..", "public", "uploads", fileName);
-
+  const fileRecord = await fileUpload.findOneAndDelete({ fileName });
+  if (!fileRecord) {
+    return next(new ErrorHandler("File not found in database.", 404));
+  }
+  // await fileUpload.deleteOne({ fileName });
   await fs.unlink(filePath, (err) => {
     if (err) {
       // If an error occurs (e.g., file not found)
@@ -55,14 +59,29 @@ exports.deleteFile = tryCatch(async (req, res, next) => {
   });
 });
 
+// retrive filse based on user ID..
 exports.getUserFiles = tryCatch(async (req, res, next) => {
   const userId = req.user?.id;
   const userFile = await fileUpload.find({ userId });
-  if (!userFile || userFile === 0) {
+  if (!userFile || userFile.length === 0) {
     return next(new ErrorHandler("No file Found for this user", 404));
   }
   res.status(200).json({
     success: true,
     files: userFile,
+  });
+});
+
+// retrive all files
+exports.getAllFiles = tryCatch(async (req, res, next) => {
+  const allFiles = await fileUpload.find();
+
+  if (!allFiles || allFiles === 0) {
+    return next(new ErrorHandler("No files found", 404));
+  }
+  res.status(200).json({
+    success: true,
+    fileCount: allFiles.length,
+    files: allFiles,
   });
 });
